% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports_rawdata.R
\name{import_infogreffes_from2014}
\alias{import_infogreffes_from2014}
\title{Import data from 2014 file}
\usage{
import_infogreffes_from2014(year, source_file)
}
\arguments{
\item{year}{}

\item{source_file}{}
}
\value{
a tibble
}
\description{
Import data from 2014 file
}
\examples{
import_infogreffes_from2014(year = 2013, source_file = "raw-data/chiffres-cles-2014.csv")
import_infogreffes_from2014(year = 2014, source_file = "raw-data/chiffres-cles-2014.csv")
import_infogreffes_from2014(year = 2015, source_file = "raw-data/chiffres-cles-2014.csv")

}

