% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports_rawdata.R
\name{import_infogreffes_from2015}
\alias{import_infogreffes_from2015}
\title{Imports data from 2015 file}
\usage{
import_infogreffes_from2015(year, source_file)
}
\arguments{
\item{year}{}

\item{source_file}{}
}
\value{
a tibble
}
\description{
Imports data from 2015 file
}
\examples{
import_infogreffes_from2015(year = 2013, source_file = "raw-data/chiffres-cles-2015.csv")
import_infogreffes_from2015(year = 2014, source_file = "raw-data/chiffres-cles-2015.csv")
import_infogreffes_from2015(year = 2015, source_file = "raw-data/chiffres-cles-2015.csv")

}

