% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports_rawdata.R
\name{import_infogreffes_from2016}
\alias{import_infogreffes_from2016}
\title{Import data from the 2016 file}
\usage{
import_infogreffes_from2016(year, source_file)
}
\arguments{
\item{year}{}

\item{source_file}{}
}
\value{
a tibble
}
\description{
Import data from the 2016 file
}
\examples{
import_infogreffes_from2016(year = 2016, source_file = "raw-data/chiffres-cles-2016.csv")

}

